requirements.txt
requests
datasette
Pillow
PyYAML